requirements.txt
djangorestframework
django-environ==0.4.5
pytest==7.*
pytest-env==0.*
django-filter
